pyproject.toml
[tool.poetry]
name = "notifications-api"
version = "0.1.0"
description = ""
authors = ["David Jetter <davidajetter@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pydantic = {version = "^1.9.0"}
python-json-logger = "^2.0.1"
requests = "^2.26.0"
boto3 = "^1.18.58"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^22.1.0"
mypy = "^0.931"
flake8 = "^4.0.1"
pytest-cov = "^3.0.0"
types-requests = "^2.25.0"
types-six = "^0.1.7"
Faker = "^11.3.0"
factory-boy = "^3.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "faker",
    "pythonjsonlogger",
    "factory",
    "boto3",
    "boto3.dynamodb.conditions"
]
ignore_missing_imports = true
